// src/app/model/Order.ts
export class Order {
    id: number;
    customerName: string;
    customerEmail: string;
    customerPhone: string;
    subtotal: number;
    tax: number;
    total: number;
    orderDate: Date;
    shippingAddress: string;
    billingAddress: string;
    area: string;
    deliveryMethod: string;
    currency: string;
    referenceNumber: string;
  
    // constructor(
    //   id: number,
    //   customerName: string,
    //   customerEmail: string,
    //   customerPhone: string,
    //   subtotal: number,
    //   tax: number,
    //   total: number,
    //   orderDate: Date,
    //   shippingAddress: string,
    //   billingAddress: string,
    //   area: string,
    //   deliveryMethod: string,
    //   currency: string,
    //   referenceNumber: string
    // ) {
    //   this.id = id;
    //   this.customerName = customerName;
    //   this.customerEmail = customerEmail;
    //   this.customerPhone = customerPhone;
    //   this.subtotal = subtotal;
    //   this.tax = tax;
    //   this.total = total;
    //   this.orderDate = orderDate;
    //   this.shippingAddress = shippingAddress;
    //   this.billingAddress = billingAddress;
    //   this.area = area;
    //   this.deliveryMethod = deliveryMethod;
    //   this.currency = currency;
    //   this.referenceNumber = referenceNumber;
    // }
  }
  